+++ blocklets/ai-studio/src/locales/en.ts
@@ -658,8 +658,10 @@ export default flat({
   agentEmptyTitle: '‌No Agent Created',
   agentEmptySubTitle: 'Click "+" create your first agent',
   back: 'Back',
-  router: 'Router Agent',
-  addRoute: 'Add Router Agent',
+  router: 'Decision',
+  addRoute: 'Add Decision',
+  noAgent: 'No Agent',
+  addAgent: 'Add Agent',
   diffRouteName:
     '"{agentName}" has the same "{routeName}" output parameter, but the "{routeName}" type is different, please modify to keep the type consistent ',
   diffRouteNameByType:
@@ -681,4 +683,7 @@ export default flat({
   chooseObject: 'Choose {object}',
   allContent: 'All Content',
   searchContent: 'Search Content',
+  profile: 'Profile',
+  avatar: 'Avatar',
+  outputFromInputPlaceholder: 'From Input "{input}"',
 });